refactor out common form save view code
